bug fix and test cases

+++ AssetManagement.Services/AssetService.cs
@@ -1,5 +1,6 @@
 using AssetManagement.Business;
 using AssetManagement.Entities;
+using AssetManagement.Exceptions;
 
 namespace AssetManagement.Services
 {
@@ -19,16 +20,28 @@ namespace AssetManagement.Services
 
         public bool UpdateAsset(Asset asset)
         {
+            if (_assetRepository.GetAssetById(asset.AssetId) == null)
+            {
+                throw new AssetNotFoundException(asset.AssetId);
+            }
             return _assetRepository.UpdateAsset(asset);
         }
 
         public bool DeleteAsset(int assetId)
         {
+            if (_assetRepository.GetAssetById(assetId) == null)
+            {
+                throw new AssetNotFoundException(assetId);
+            }
             return _assetRepository.DeleteAsset(assetId);
         }
 
         public Asset GetAssetById(int assetId)
         {
+            if (_assetRepository.GetAssetById(assetId) == null)
+            {
+                throw new AssetNotFoundException(assetId);
+            }
             return _assetRepository.GetAssetById(assetId);
         }
